Remove the GLOB_BRACE flag
When calling glob() to get the configuration file location, the
GLOB_BRACE flag is used. This flag enables the parsing of '{a,b}', so
that '/path/{a,b}' represents '/path/a' and '/path/b'. Looking at
Snuffleupagus' documentation I saw nowhere this kind of path was
supported. Instead the documentation shows that glob() is used for paths
with wildcards.

The use of GLOB_BRACE is problematic as it is glibc specific and is not
supported by some other C libraries, such as Musl. Snuffleupagus cannot
be used in Apline for this reason.

Since the documentation does not give a valid usage of GLOB_BRACE this
patch removes it, which fixes the non-glibc support.

Signed-off-by: Antoine Tenart <antoine.tenart@ack.tf>