Make the testsuite work on php7.4
Before php7.4, it seems that the curl module was loaded by default, but since
it's no the case anymore, it has to be manually specified in the testsuite.
Interestingly, Php's testsuite mechanism is running snippets to determine some
runtime parameters like the extension directory. Unfortunately,it tries to run
them with Snuffleupagus loaded, resulting in an error, since no configuration
file is passed.